magento/graphql-ce#1054: Add e-mail address validation to customer data validation for GraphQL
